fix : 수정 사항 반영
- 업로드 안되는 오류 해결
