+++ client/src/pages/Posting/Components/Content.jsx
@@ -1,5 +1,26 @@
+import { FileUpload } from "../../../modules/FileUpload";
+
 export const Content = ({ props }) => {
-    const { no, data, update, nextStep, prevStep, addStep } = props;
+    const { no, data, updateData, nextStep, prevStep, addStep } = props;
+
+    const update = (e) => {
+        // check if file is uploaded
+        if (e.target.files) {
+            let url = FileUpload(e.target.files[0]);
+            console.log("url: ", url);
+            updateData({
+                ...data,
+                [e.target.name]: url
+            });
+            return;
+        }
+
+        updateData({
+            ...data,
+            [e.target.name]: e.target.value
+        });
+    };
+
     return (
         <div>
             <div className="content-header">
@@ -15,9 +36,9 @@ export const Content = ({ props }) => {
                     className="input"
                     name="img_url"
                     type="file"
-                    value={data.img_url}
                     onChange={update}
                 />
+                <img src={data.img_url} alt="이미지 미리보기" />
             </div>
 
             <div className="content-body">
