+++ client/src/pages/Posting/Components/Footer.jsx
@@ -1,5 +1,24 @@
+import { FileUpload } from "../../../modules/FileUpload";
+
 export const Footer = ({ props, save }) => {
-    const { data, prevStep, update } = props;
+    const { data, prevStep, updateData } = props;
+    const update = (e) => {
+        // check if file is uploaded
+        if (e.target.files) {
+            let url = FileUpload(e.target.files[0]);
+            console.log("url: ", url);
+            updateData({
+                ...data,
+                [e.target.name]: url
+            });
+            return;
+        }
+
+        updateData({
+            ...data,
+            [e.target.name]: e.target.value
+        });
+    };
 
     return (
         <div className="footer">
