+++ client/src/pages/Posting/Posting.jsx
@@ -1,6 +1,5 @@
 import { useState } from "react";
 import axios from "axios";
-import { FileUpload } from "../../modules/FileUpload";
 
 import { Content } from "./Components/Content";
 import { Title } from "./Components/Title";
@@ -121,25 +120,6 @@ export const Posting = () => {
             });
     };
 
-    //
-    const update = (e) => {
-        // check if file is uploaded
-        if (e.target.files) {
-            let url = FileUpload(e.target.files[0]);
-            console.log("url: ", url);
-            updateData({
-                ...data,
-                [e.target.name]: url
-            });
-            return;
-        }
-
-        updateData({
-            ...data,
-            [e.target.name]: e.target.value
-        });
-    };
-
     const loadComponent = () => {
         const props = {
             data: data[step].data,
@@ -147,7 +127,9 @@ export const Posting = () => {
             nextStep,
             prevStep,
             addStep,
-            update
+            updateData: (d) => {
+                updateData(step, d);
+            }
         };
 
         switch (data[step].type) {
